mock_app/requirements.txt
biopython==1.73
Click==7.0
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.1
pyarrow==0.12.0
pandas==0.24.1
pyparsing==2.3.1
pysam==0.15.2
python-dateutil==2.8.0
pytz==2018.9
scipy==1.2.0
seaborn==0.9.0
six==1.12.0
